base class first approach
